Rename app module to polebot
