+++ .vscode/launch.json
@@ -5,10 +5,10 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Python Debugger: Module",
+            "name": "Python Debug Polebot",
             "type": "debugpy",
             "request": "launch",
-            "module": "app"
+            "module": "polebot"
         }
     ]
 }
