--- src/app/__init__.py
@@ -1,1 +1,0 @@
-"""The app module, containing the application's code."""
