+++ src/polebot/__init__.py
@@ -1,0 +1,1 @@
+"""The polebot module, containing the application's code."""
