+++ src/polebot/exceptions.py
