+++ src/polebot/map_selector/__init__.py
