+++ src/polebot/map_selector/config_loader.py
