+++ src/polebot/map_selector/selector.py
