+++ src/polebot/votemap_manager.py
