+++ tests/polebot/test_converters/get_maps.json
