+++ tests/polebot/test_converters/get_votemap_config.json
