+++ tests/polebot/test_converters/set_votemap_whitelist.json
