+++ tests/polebot/test_map_selector/get_maps.json
