+++ tests/polebot/test_map_selector/server_config.json
