+++ tests/polebot/test_votemap_manager/get_maps.json
